debug soil_parm_ID and change beetle default input type

+++ rhessys/include/rhessys.h
@@ -3274,14 +3274,14 @@ struct patch_beetle_object
 struct beetle_default {
      int ID;
      double attack_mortality;
-     int year_delay; //snag pool delay
-     int half_life;
+     double year_delay; //snag pool delay
+     double half_life;
      int n_rows;
      int n_cols;
      int beetle_in_buffer;
      int year_attack;
-     int leaf_year_delay; //red needle delay uear
-     int leaf_half_life;  // red needle half life
+     double leaf_year_delay; //red needle delay uear
+     double leaf_half_life;  // red needle half life
      int calc_single_attack;
      double min_abc; // the minmum above ground carbon that a patch have to make sure an attack happen g/m2
      int mortality_type; //type 1 is beetle attack, type 2 is fire, type 3 is thinning/Ning Ren 2019/04/30
